Initial commit of symbols as separate repository
This is a set of symbols released to the public domain by
Norman Megill to support the Metamath project. It was originally
posted on the metamath website here:

<http://us.metamath.org/downloads/symbols.tar.bz2>

We are in the process of transitioning to check all source files
into a repostiroy, and then generating the website *from* them.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>